# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Manga Collection API",
        "description": "Manga Collection API Docs",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "local server"
        },
        {
            "url": "https://pink-buffalo-749567.hostingersite.com/api",
            "description": "production server"
        }
    ],
    "paths": {
        "/authors": {
            "get": {
                "tags": [
                    "Authors"
                ],
                "summary": "Index",
                "operationId": "c0d797ed63303dad3c42bee5c36cf40c",
                "parameters": [
                    {
                        "name": "ordering",
                        "in": "query",
                        "description": "Parameter to order results",
                        "required": false
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Parameter that set the page",
                        "required": false
                    },
                    {
                        "name": "page_size",
                        "in": "query",
                        "description": "Parameter that set the size of the result collection",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of authors"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Authors"
                ],
                "summary": "Store",
                "operationId": "a05b6d79ca79b1d32c3651f1ced437b7",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Akira Amano"
                                    },
                                    "picture": {
                                        "description": "Must be a base64 encoded image",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Author created successfully"
                    },
                    "422": {
                        "description": "Field Error"
                    },
                    "401": {
                        "description": "Not allowed"
                    },
                    "499": {
                        "description": "Not allowed"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/authors/{author_id}": {
            "get": {
                "tags": [
                    "Authors"
                ],
                "summary": "Show",
                "operationId": "499b4c845a260cf293ffbfbf6fd4fb26",
                "parameters": [
                    {
                        "name": "author_id",
                        "in": "path",
                        "description": "Parameter that filter the entity",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return the requested author"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Not allowed"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Authors"
                ],
                "summary": "Destroy",
                "operationId": "f0282542ac478412cc839b612d25b977",
                "parameters": [
                    {
                        "name": "author_id",
                        "in": "path",
                        "description": "Parameter that filter the entity",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Author deleted successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Not allowed"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Authors"
                ],
                "summary": "Update",
                "operationId": "ba5333c5d624bc5c6846fbd41b22a2ac",
                "parameters": [
                    {
                        "name": "author_id",
                        "in": "path",
                        "description": "Parameter that filter the entity",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Osamu Tezuka"
                                    },
                                    "picture": {
                                        "description": "Must be a base64 encoded image",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Author updated successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Not allowed"
                    },
                    "499": {
                        "description": "Not allowed"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "500": {
                        "description": "Not allowed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}